# Conflicts:
#	.idea/caches/build_file_checksums.ser
#	app/src/main/java/com/example/igor/rugetsahackathon2018/Edge.java
#	app/src/main/java/com/example/igor/rugetsahackathon2018/MainActivity.java
#	app/src/main/java/com/example/igor/rugetsahackathon2018/MapsActivity.java
#	app/src/main/java/com/example/igor/rugetsahackathon2018/Node.java

+++ app/src/main/java/com/example/igor/rugetsahackathon2018/Node.java
@@ -35,7 +35,7 @@ public class Node implements Serializable {
 
     }
 
-    public Node(String name, float latitude, float longitude) {
+    public Node(String name, double latitude, double longitude) {
         this.name = name;
         this.latitude = latitude;
         this.longitude = longitude;
